+++ ingredient/ingredient.c
@@ -2,7 +2,6 @@
 // Created by dawids21 on 4/28/21.
 //
 
-#include <array.h>
 #include "ingredient.h"
 
 typedef enum {
@@ -16,17 +15,3 @@ struct Ingredient {
     int amount;
     IngredientType type;
 };
-
-struct IngredientService {
-    Array ingredients;
-};
-
-IngredientService new_ingredient_service() {
-    IngredientService service = (IngredientService) malloc(sizeof(struct IngredientService));
-    service->ingredients = new_array();
-    return service;
-}
-
-void delete_ingredient_service(IngredientService service) {
-    free(service);
-}
